refactor calculate #62
