+++ snugh/lecture/tests.py
@@ -4,7 +4,7 @@ from rest_framework import status
 from pathlib import Path
 from django.db.models import Q
 from user.utils import UserFactory, UserMajorFactory
-from .utils import SemesterFactory, SemesterLectureFactory
+from .utils_test import SemesterFactory, SemesterLectureFactory
 from .models import Lecture, Plan, PlanMajor, Semester, SemesterLecture
 
 BASE_DIR = Path(__file__).resolve().parent.parent
